backend/requirements.txt
Flask>=2.0
flask-socketio>=5.0
python-dotenv
SQLAlchemy>=1.4
psycopg2-binary
alembic
ariadne>=0.17
flask-cors
azure-storage-blob
python-jose  
eventlet  
redis  
bcrypt
